refactor: put client transport dispatch into conn module
